DownloadCore: Fix some bugs.

+++ SeaMinecraftLauncherCore/Tools/DownloadCore.cs
@@ -92,17 +92,20 @@ namespace SeaMinecraftLauncherCore.Tools
 
         public static event EventHandler<DownloadSuccessEventArgs> DownloadSuccess;
 
-        public static List<Task> TryDownloadFiles(IEnumerable<DownloadInfo> downInfos, int retryCount = 5, int timeout = 2000)
+        public static async Task<List<Task>> TryDownloadFiles(IEnumerable<DownloadInfo> downInfos, int retryCount = int.MaxValue, int threadCount = int.MaxValue, int timeout = 15000)
         {
             DownloadSuccess += delegate { };
             List<Task> asyncPool = new List<Task>();
             int failedCount = 0;
+            int openedThreadCount = 0;
             foreach (DownloadInfo downInfo in downInfos)
             {
                 asyncPool.Add(Task.Run(async () =>
                 {
+                    while (openedThreadCount >= threadCount) { }
+                    openedThreadCount++;
                     bool result = true;
-                    Console.WriteLine($"开始下载 {downInfo.Url}");
+                    //Console.WriteLine($"开始下载 {downInfo.Url}");
                     for (int i = 0; i < (retryCount == int.MaxValue ? retryCount : retryCount + 1); i++)
                     {
                         if (await TryDownloadFileAsync(downInfo, timeout))
@@ -114,13 +117,15 @@ namespace SeaMinecraftLauncherCore.Tools
                     result = false;
                 SkipRetry:
                     DownloadSuccess(null, new DownloadSuccessEventArgs(downInfo, result));
+                    openedThreadCount--;
                 }));
             }
 
+            //await Task.WhenAll(asyncPool);
             return asyncPool;
         }
 
-        public static async Task<bool> TryDownloadFileAsync(DownloadInfo downInfo, int timeout = 2000)
+        public static async Task<bool> TryDownloadFileAsync(DownloadInfo downInfo, int timeout = 10000)
         {
             using (var webClient = new WebClientPlus(timeout))
             {
@@ -139,12 +144,20 @@ namespace SeaMinecraftLauncherCore.Tools
                 {
                     result = e.Error == null ? true : false;
                 };
-                await webClient.DownloadFileTaskAsync(downInfo.Url, Path.Combine(downInfo.DownloadPath, fileNameWithOtherExt));
-                if (File.Exists(Path.Combine(downInfo.DownloadPath, fileName)))
+                var downTask = webClient.DownloadFileTaskAsync(downInfo.Url, Path.Combine(downInfo.DownloadPath, fileNameWithOtherExt));
+                await Task.Delay(timeout);
+                if (downTask.IsCompleted)
                 {
-                    File.Delete(Path.Combine(downInfo.DownloadPath, fileName));
+                    if (File.Exists(Path.Combine(downInfo.DownloadPath, fileName)))
+                    {
+                        File.Delete(Path.Combine(downInfo.DownloadPath, fileName));
+                    }
+                    FileSystem.RenameFile(Path.Combine(downInfo.DownloadPath, fileNameWithOtherExt), fileName);
                 }
-                FileSystem.RenameFile(Path.Combine(downInfo.DownloadPath, fileNameWithOtherExt), fileName);
+                else
+                {
+                    result = false;
+                }
                 return result;
             }
         }
